requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.10.5
django-braces==1.11.0
django-extensions==1.7.6
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest-framework-social-oauth2==1.0.5
djangorestframework==3.5.4
enum34==1.1.6
httpie==0.9.9
interview==1.0
ipython==5.2.2
ipython-genutils==0.1.0
Markdown==2.6.8
oauthlib==1.1.2
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
py==1.4.32
Pygments==2.2.0
PyJWT==1.4.2
pytest==3.0.5
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
traitlets==4.3.1
wcwidth==0.1.7
